refactor(MenuGroup at el): after separate packages, move entity, service, etc about MenuGroup
